requirements.txt
scipy>=1.10
timm>=1.0.3
grad-cam>=1.4.0
shap>=0.43
torchmetrics>=1.0
pandas>=1.5
numpy>=1.23
scikit-learn>=1.2
tqdm>=4.65
Pillow>=9.0
